backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
flasgger==0.9.7.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
marshmallow==3.21.3
scikit-learn>=1.6,<2
pandas>=2.2.2,<3
numpy>=1.26.4,<3
joblib==1.4.2
yfinance==0.2.54
python-dotenv==1.0.1
pytest==8.3.2
pytest-cov==5.0.0
dotenv